Feature/scenario scale kubernetes cluster (#138)
* Added GetServerNamesAction

* Added DetachNodeFromKubernetesClusterAction

* Modified DeleteInstanceAction such that a namePattern can be passed

* Async DroolsService

* Stateful CronChecker

* Bumped VER to 0.0.3.6

+++ src/main/java/org/k8loud/executor/actions/openstack/DeleteInstanceAction.java
@@ -1,11 +1,13 @@
 package org.k8loud.executor.actions.openstack;
 
-import lombok.EqualsAndHashCode;
-import org.k8loud.executor.model.Params;
 import lombok.Builder;
+import lombok.EqualsAndHashCode;
+import org.apache.commons.collections4.CollectionUtils;
 import org.k8loud.executor.exception.ActionException;
 import org.k8loud.executor.exception.OpenstackException;
+import org.k8loud.executor.exception.ParamNotFoundException;
 import org.k8loud.executor.exception.ValidationException;
+import org.k8loud.executor.model.Params;
 import org.k8loud.executor.openstack.OpenstackService;
 
 import java.util.Collections;
@@ -15,7 +17,7 @@ import java.util.Map;
 @EqualsAndHashCode
 public class DeleteInstanceAction extends OpenstackAction {
     private String region;
-    private String name;
+    private String namePattern;
     private List<String> serverIds;
 
     public DeleteInstanceAction(Params params, OpenstackService openstackService) throws ActionException {
@@ -23,26 +25,30 @@ public class DeleteInstanceAction extends OpenstackAction {
     }
 
     @Builder
-    public DeleteInstanceAction(OpenstackService openstackService, String region, String name, List<String> serverIds) {
+    public DeleteInstanceAction(OpenstackService openstackService, String region, String namePattern,
+                                List<String> serverIds) {
         super(openstackService);
         this.region = region;
-        this.name = name;
+        this.namePattern = namePattern;
         this.serverIds = serverIds;
     }
 
     @Override
     public void unpackParams(Params params) {
         region = params.getRequiredParam("region");
-        name = params.getRequiredParam("name");
+        namePattern = params.getOptionalParam("namePattern", null);
         serverIds = params.getOptionalParamAsListOfStrings("serverIds", Collections.emptyList());
     }
 
     @Override
     protected Map<String, String> executeBody() throws OpenstackException, ValidationException {
-        if (serverIds.isEmpty()){
-            return openstackService.deleteServers(region, name);
+        final boolean serverIdsIsEmpty = CollectionUtils.isEmpty(serverIds);
+        if (namePattern == null && serverIdsIsEmpty) {
+            throw new ParamNotFoundException("Either namePattern or serverIds should be provided");
         }
-
+        if (serverIdsIsEmpty) {
+            return openstackService.deleteServers(region, namePattern);
+        }
         return openstackService.deleteServers(region, serverIds);
     }
 }
